parser - move Nmea to own module
